+++ modules/winwave/src.c
@@ -160,7 +160,7 @@ static int read_stream_open(struct ausrc_st *st, const struct ausrc_prm *prm,
 	wfmt.nChannels       = prm->ch;
 	wfmt.nSamplesPerSec  = prm->srate;
 	wfmt.wBitsPerSample  = (WORD)(st->sampsz * 8);
-	wfmt.nBlockAlign     = prm->ch * st->sampsz;
+	wfmt.nBlockAlign     = (WORD)(prm->ch * st->sampsz);
 	wfmt.nAvgBytesPerSec = wfmt.nSamplesPerSec * wfmt.nBlockAlign;
 	wfmt.cbSize          = 0;
